requirements.txt
Flask
tweepy==3.10
requests
SQLAlchemy
SQLAlchemy-Utils
psycopg2
schedule
Pillow
python-dotenv
pyyaml
